Get the fit to obtain the clocks posteriors

Obtain the approximate posterior of the clocks for each model fit with the number of components up to k_max.

Usage

fit_h(
  x,
  max_attempts = 2,
  INIT = TRUE,
  tolerance = 0.01,
  possible_k = c("2:1", "2:2", "2:0"),
  alpha = 0.05,
  min_mutations_number = 2
)

Arguments

x

list: A CNAqc object.

max_attempts

num: max number of repeated inference for ADVI

INIT

logical: boolean variable to set the initialization phase to TRUE or FALSE

tolerance

num: tolerance in the ELBO optimization procedure

possible_k

chr: "2:1" "2:2" "2:0"

alpha

num: (type double) confidence interval level to choose the data that fall in the expected binomial intervals

min_mutations_number

num: (type double) minimum number of accepted mutations for a segment to be included in the inference

Value

results_and_data = list(data = input_data_list, results = results, output_files_list = output_files_list)